Fixed package tests

+++ convert/convert_test.go
@@ -9,26 +9,26 @@ import (
 )
 
 func TestDecToInt(t *testing.T) {
-	assert.Equal(t, 0, convert.DecToInt([]byte("")))
-	assert.Equal(t, 0, convert.DecToInt([]byte("0")))
-	assert.Equal(t, 1, convert.DecToInt([]byte("1")))
-	assert.Equal(t, 10, convert.DecToInt([]byte("10")))
-	assert.Equal(t, 100, convert.DecToInt([]byte("100")))
-	assert.Equal(t, 123456789, convert.DecToInt([]byte("123456789")))
-	assert.Equal(t, 0, convert.DecToInt([]byte("ZZZ")))
+	assert.Equal(t, int64(0), convert.DecToInt([]byte("")))
+	assert.Equal(t, int64(0), convert.DecToInt([]byte("0")))
+	assert.Equal(t, int64(1), convert.DecToInt([]byte("1")))
+	assert.Equal(t, int64(10), convert.DecToInt([]byte("10")))
+	assert.Equal(t, int64(100), convert.DecToInt([]byte("100")))
+	assert.Equal(t, int64(123456789), convert.DecToInt([]byte("123456789")))
+	assert.Equal(t, int64(0), convert.DecToInt([]byte("ZZZ")))
 }
 
 func TestHexToInt(t *testing.T) {
-	assert.Equal(t, 0x0, convert.HexToInt([]byte("")))
-	assert.Equal(t, 0x0, convert.HexToInt([]byte("0")))
-	assert.Equal(t, 0x1, convert.HexToInt([]byte("1")))
-	assert.Equal(t, 0xA, convert.HexToInt([]byte("A")))
-	assert.Equal(t, 0x10, convert.HexToInt([]byte("10")))
-	assert.Equal(t, 0xAFFE, convert.HexToInt([]byte("AFFE")))
-	assert.Equal(t, 0xAFFE, convert.HexToInt([]byte("Affe")))
-	assert.Equal(t, 0xBADFACE, convert.HexToInt([]byte("BADFACE")))
-	assert.Equal(t, 0xBADFACE, convert.HexToInt([]byte("BadFace")))
-	assert.Equal(t, 0, convert.HexToInt([]byte("ZZZ")))
+	assert.Equal(t, int64(0x0), convert.HexToInt([]byte("")))
+	assert.Equal(t, int64(0x0), convert.HexToInt([]byte("0")))
+	assert.Equal(t, int64(0x1), convert.HexToInt([]byte("1")))
+	assert.Equal(t, int64(0xA), convert.HexToInt([]byte("A")))
+	assert.Equal(t, int64(0x10), convert.HexToInt([]byte("10")))
+	assert.Equal(t, int64(0xAFFE), convert.HexToInt([]byte("AFFE")))
+	assert.Equal(t, int64(0xAFFE), convert.HexToInt([]byte("Affe")))
+	assert.Equal(t, int64(0xBADFACE), convert.HexToInt([]byte("BADFACE")))
+	assert.Equal(t, int64(0xBADFACE), convert.HexToInt([]byte("BadFace")))
+	assert.Equal(t, int64(0), convert.HexToInt([]byte("ZZZ")))
 }
 
 func BenchmarkDecToInt(b *testing.B) {
